requirements.txt
paho-mqtt>=2.1.0
cryptography>=44.0.2
meshtastic>=2.6.1
pyaml>=25.1.0
sqlalchemy>=2.0.40
psycopg2>=2.9.10